Change name to Aphelia
Development name was MyBit; change all references of MyBit/Bit to Aphelia
