+++ src/main/java/to/us/awesomest/aphelia/module/chathandlers/ChatHandler.java
@@ -1,0 +1,8 @@
+package to.us.awesomest.aphelia.module.chathandlers;
+
+import to.us.awesomest.aphelia.module.Module;
+
+
+public interface ChatHandler extends Module {
+    boolean adminException();
+}
