+++ src/main/java/to/us/awesomest/aphelia/module/command/Command.java
@@ -1,0 +1,7 @@
+package to.us.awesomest.aphelia.module.command;
+
+import to.us.awesomest.aphelia.module.Module;
+
+public interface Command extends Module {
+    boolean isDMUsable();
+}
