+++ src/main/java/to/us/awesomest/aphelia/module/command/CommandUtils.java
@@ -1,0 +1,7 @@
+package to.us.awesomest.aphelia.module.command;
+
+final class CommandUtils {
+    static String[] parseArgs(String argString) {
+        return argString.split(" ");
+    }
+}
