--- src/main/java/to/us/awesomest/mybit/module/command/CheckComs.java
@@ -1,31 +1,0 @@
-package to.us.awesomest.mybit.module.command;
-
-import net.dv8tion.jda.api.entities.Guild;
-import net.dv8tion.jda.api.entities.MessageChannel;
-import net.dv8tion.jda.api.entities.User;
-import org.slf4j.LoggerFactory;
-
-public class CheckComs implements Command {
-    //Implement singleton handling
-    static CheckComs instance = null;
-
-    @Override
-    public void run(User author, MessageChannel channel, String command, Guild guild) {
-        channel.sendMessage("ComChecks complete, standby for launch.").queue();
-    }
-
-    @Override
-    public String getName() {
-        return "checkComs";
-    }
-
-    @Override
-    public String getDescription() {
-        return "Check if all of the communications with your shard are working. As if you were launching a rocket.";
-    }
-
-    @Override
-    public boolean isDMUsable() {
-        return true;
-    }
-}
