--- src/main/java/to/us/awesomest/mybit/module/command/CommandUtils.java
@@ -1,7 +1,0 @@
-package to.us.awesomest.mybit.module.command;
-
-final class CommandUtils {
-    static String[] parseArgs(String argString) {
-        return argString.split(" ");
-    }
-}
